requirements.txt
python-telegram-bot==21.3
python-dotenv==1.0.1
google-api-python-client==2.104.0
sentence-transformers==2.5.1
PyPDF2==3.0.1
python-docx==1.1.2
pytesseract==0.3.10
Pillow==10.2.0
beautifulsoup4==4.12.3
reportlab==4.0.8
requests==2.31.0
torch==2.3.0